tools/jd_responsibility_extractor/requirements.txt
strands-agents>=0.2.0
strands-agents-tools>=0.2.0
requests>=2.31.0
pydantic>=2.6.0,<3
pydantic-core>=2.16.0,<3
# boto3 is provided in the AWS Lambda runtime. If building outside Lambda or USE_RUNTIME_BOTO3=false, uncomment:
# boto3>=1.34.0

